+++ build-logic/convention/src/main/kotlin/project/convention/logic/config/LibraryPluginConfig.kt
@@ -33,9 +33,7 @@ enum class LibraryModule(val path: String) {
     CommonFeature(":common-feature"),
     StartupFeature(":startup-feature"),
     OnboardingFeature(":onboarding-feature"),
-    DashboardFeature(":dashboard-feature"),
     PresentationFeature(":presentation-feature"),
-    ProximityFeature(":proximity-feature"),
     IssuanceFeature(":issuance-feature"),
     LandingFeature(":landing-feature");